Don't allow for client.destroy to be overwritten.
Prior to this it was possible for someone to set client.destroy to be a
function including something rather nasty, e.g. a while(true), loop that would not be
protected by the vm module.  This would leave the bot in a state of
unresponsiveness that couldn't be fixed by anyone without access to the
repl the bot was running on.